+++ standalone-metastore/metastore-server/src/main/java/org/apache/hadoop/hive/metastore/RetryingHMSHandler.java
@@ -178,6 +178,6 @@ public class RetryingHMSHandler extends AbstractHMSHandlerProxy {
     if (!(t instanceof JDOException || t instanceof NucleusException)) {
       return false;
     }
-    return RecoverableExceptions.isRecoverableException(t);
+    return DatabaseProduct.isRecoverableException(t);
   }
 }
